refactor: realocated screens to PageTextAnimated
