+++ src/components/PageTextAnimated/Coincidence/Coincidence.styles.ts
