+++ src/components/PageTextAnimated/Coincidence/Coincidence.tsx
