+++ src/components/PageTextAnimated/Coincidence/index.tsx
