+++ src/components/PageTextAnimated/Coincidence/interfaces/index.ts
