+++ src/components/PageTextAnimated/Home/Home.styles.ts
