+++ src/components/PageTextAnimated/Home/Home.tsx
