+++ src/components/PageTextAnimated/Home/index.tsx
