+++ src/components/PageTextAnimated/Home/interfaces/index.ts
