+++ src/components/PageTextAnimated/Potency/Potency.tsx
