+++ src/components/PageTextAnimated/Potency/index.tsx
