+++ src/components/PageTextAnimated/Potency/interfaces/index.ts
