+++ src/components/PageTextAnimated/Potency/styles.ts
