+++ src/components/PageTextAnimated/SeeMore/styles.ts
