+++ src/components/PageTextAnimated/TimeMachine/TimeMachine.tsx
