+++ src/components/PageTextAnimated/TimeMachine/index.tsx
