+++ src/components/PageTextAnimated/TimeMachine/interfaces/index.ts
