+++ src/components/PageTextAnimated/TimeMachine/styles.ts
