+++ src/components/PageTextAnimated/Universe/Universe.styles.ts
