+++ src/components/PageTextAnimated/Universe/Universe.tsx
