+++ src/components/PageTextAnimated/Universe/interfaces/index.ts
