+++ src/components/PageTextAnimated/common/interfaces/index.ts
