+++ src/components/PageTextAnimated/common/styles/index.ts
